// src/model/StationDetail.ts
export interface StationDetail {
  success: boolean;
  code: string;
  msg: string;
  data: Data;
}

export interface Data {
  id: string;
  createDate: number;
  updateDate: number;
  deleteFlag: number;
  simFlowState: number;
  connectTime: number;
  stationName: string;
  dataTimestamp: string;
  userId: string;
  userName: string;
  userEmail: string;
  orgCode: string;
  visitorCount: number;
  addr: string;
  timeZone: number;
  timeZoneStr: string;
  timeZoneId: string;
  timeZoneName: string;
  daylight: number;
  mpptSwitch: number;
  fullHour: number;
  fullHourStr: string;
  monthCarbonDioxide: number;
  capacity: number;
  capacityStr: string;
  capacityPec: string;
  dip: number;
  dipStr: string;
  azimuth: number;
  azimuthStr: string;
  money: string;
  price: number;
  installer: string;
  installerId: string;
  dateTime: string;
  offset: number;
  offsetStr: string;
  longitude: string;
  latitude: string;
  state: number;
  dayEnergy: number;
  dayEnergyStr: string;
  dayInCome: number;
  dayInComeUnit: string;
  monthEnergy: number;
  monthEnergyStr: string;
  monthInCome: number;
  monthInComeUnit: string;
  yearEnergy: number;
  yearEnergyStr: string;
  yearInCome: number;
  yearInComeUnit: string;
  allInCome: number;
  allInComeUnit: string;
  allEnergy: number;
  allEnergy1: number;
  allEnergyStr: string;
  powerStationNumTree: number;
  powerStationNumTreeUnit: string;
  powerStationAvoidedCo2: number;
  powerStationAvoidedCo2Unit: string;
  power: number;
  powerStr: string;
  powerPec: string;
  porwerPercent: number;
  batteryPower: number;
  batteryPowerStr: string;
  batteryPowerPec: string;
  batteryPowerZheng: number;
  batteryPowerFu: number;
  batteryPercent: number;
  familyLoadPercent: number;
  psum: number;
  psumStr: string;
  psumPec: string;
  psumZheng: number;
  psumFu: number;
  oneSelf: number;
  gridPurchasedTotalEnergy: number;
  gridPurchasedTotalEnergyStr: string;
  gridSellTotalEnergy: number;
  gridSellTotalEnergyStr: string;
  gridPurchasedEnergy: number;
  gridPurchasedEnergyStr: string;
  gridSellEnergy: number;
  gridSellEnergyStr: string;
  gridPurchasedDayEnergy: number;
  gridPurchasedDayEnergyStr: string;
  gridSellDayEnergy: number;
  gridSellDayEnergyStr: string;
  gridPurchasedMonthEnergy: number;
  gridPurchasedMonthEnergyStr: string;
  gridSellMonthEnergyStr: string;
  gridPurchasedYearEnergy: number;
  gridPurchasedYearEnergyStr: string;
  gridSellYearEnergyStr: string;
  batteryDischargeEnergy: number;
  batteryDischargeEnergyStr: string;
  batteryChargeEnergy: number;
  batteryChargeEnergyStr: string;
  batteryDischargeMonthEnergy: number;
  batteryDischargeMonthEnergyStr: string;
  batteryChargeMonthEnergy: number;
  batteryDischargeYearEnergy: number;
  batteryDischargeYearEnergyStr: string;
  batteryChargeYearEnergy: number;
  batteryDischargeTotalEnergy: number;
  batteryDischargeTotalEnergyStr: string;
  batteryChargeTotalEnergy: number;
  familyLoadPower: number;
  familyLoadPowerStr: string;
  familyLoadPowerPec: string;
  homeGridTodayEnergy: number;
  homeGridTodayEnergyStr: string;
  homeGridMonthEnergy: number;
  homeGridMonthEnergyStr: string;
  homeGridYearEnergy: number;
  homeGridYearEnergyStr: string;
  homeGridTotalEnergy: number;
  homeGridTotalEnergyStr: string;
  backupTodayEnergy: number;
  backupTodayEnergyStr: string;
  backupMonthEnergy: number;
  backupMonthEnergyStr: string;
  backupYearEnergy: number;
  backupYearEnergyStr: string;
  backupTotalEnergy: number;
  backupTotalEnergyStr: string;
  totalLoadPower: number;
  totalLoadPowerStr: string;
  bypassLoadPower: number;
  bypassLoadPowerStr: string;
  homeLoadEnergy: number;
  homeLoadEnergyStr: string;
  homeLoadMonthEnergy: number;
  homeLoadYearEnergy: number;
  homeLoadTotalEnergy: number;
  picUrl: string;
  countryStr: string;
  regionStr: string;
  cityStr: string;
  countyStr: string;
  weather: string;
  sr: string;
  ss: string;
  tmpMax: string;
  tmpMin: string;
  tmpUnit: string;
  condTxtD: string;
  condTxtN: string;
  condCodeD: string;
  condCodeN: string;
  hum: string;
  weatherUpdateDate: string;
  pcpn: string;
  windSpd: string;
  windDir: string;
  sno: string;
  weatherType: number;
  windSpeed: number;
  windDirection: number;
  humidity: number;
  temp: number;
  rainfall: number;
  airPressure: number;
  contribution: number;
  synchronizationType: number;
  epmType: number;
  gridSwitch: number;
  type: number;
  stationTypeNew: number;
  screenMap: number;
  screenGuideState: number;
  storedInverterType: number;
  countryShortName: string;
  inverterPower: number;
  fisGenerateTime: number;
  homeLoadTotalEnergyStr: string;
  homeLoadMonthEnergyStr: string;
  homeLoadYearEnergyStr: string;
  gridMonthEnergy: number;
  gridYearEnergy: number;
  inverterPowerStr: string;
  batteryChargeMonthEnergyStr: string;
  batteryChargeYearEnergyStr: string;
  batteryChargeTotalEnergyStr: string;
}